Moved AsyncReadBlock to blockreader
